pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Igor Botsii <igorbotsii42@gmail.com>"]
package-mode = false


[tool.poetry.dependencies]
python = ">=3.14"
Django = "^6.0.3"
djangorestframework = "^3.17.1"
mysqlclient = "^2.2.8"
django-filter = "^25.2"
djangorestframework-simplejwt = "^5.5.1"
timedelta = "^2020.12.3"
pillow = "^12.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "isort (>=8.0.1,<9.0.0)"
]
